test(code-qualities): move the failure-reporting tests to their own module
Appending them pushed `test_assess_regression.py` from 339 to 382 lines, past
the taste-lints size warning at 375, which the taste count ratchet caught as a
regression from baseline 575.

They belong apart anyway: that module covers the regression gate of issue
#4364, while these cover whether a failed file is reported at all. The
regression module is back to 339 lines and the ratchet is back at baseline.

Refs #4981
