+++ tests/test_assess_regression.py
@@ -38,7 +38,6 @@ _SKILL_SCRIPTS = (
 sys.path.insert(0, str(_SKILL_SCRIPTS))
 
 # ruff: noqa: E402
-import assess
 from assess import (
     ChangedFile,
     FileAssessment,
@@ -338,45 +337,3 @@ class TestGetBaseAssessments:
 
         assert str(file_path) in result
         assert result[str(file_path)].file_path == str(file_path)
-
-
-class TestAssessFilesReportsFailures:
-    """A file that fails to assess is reported, never dropped.
-
-    Dropping it made the failure invisible: `summary.file_count` is the length
-    of the returned list, so a run that lost four of five files still looked
-    internally consistent and a consumer could read the survivor as evidence
-    that the whole change was assessed.
-    """
-
-    def test_failed_file_is_recorded_as_unscored(self, tmp_path):
-        good = tmp_path / "good.py"
-        good.write_text("def f():\n    return 1\n", encoding="utf-8")
-        bad = tmp_path / "bad.py"
-        bad.write_text("def g():\n    return 2\n", encoding="utf-8")
-
-        real = assess.assess_file
-
-        def _explode(path, context, use_serena):
-            if path == bad:
-                raise RuntimeError("boom")
-            return real(path, context, use_serena)
-
-        with patch.object(assess, "assess_file", side_effect=_explode):
-            results = assess._assess_files([good, bad], "production", False)
-
-        assert len(results) == 2, "the failed file must still appear in the list"
-        by_path = {r.file_path: r for r in results}
-        failed = by_path[str(bad)]
-        assert failed.cohesion.confidence == 0.0
-        assert any("assessment failed" in r for r in failed.cohesion.reasons)
-        assert by_path[str(good)].cohesion.confidence > 0.0
-
-    def test_all_files_failing_yields_all_unscored(self, tmp_path):
-        """Negative control: nothing is silently salvaged."""
-        f = tmp_path / "a.py"
-        f.write_text("def f():\n    return 1\n", encoding="utf-8")
-        with patch.object(assess, "assess_file", side_effect=RuntimeError("boom")):
-            results = assess._assess_files([f], "production", False)
-        assert len(results) == 1
-        assert results[0].cohesion.confidence == 0.0
